+++ client/src/pages/cart/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './styles.less'
+
+function Cart (props) {
+  return (
+    <div>Cart Page</div>
+  )
+}
+
+export default Cart;
